Load data from remote for sleep quality trend view.

+++ app/src/main/java/im/hch/sleeprecord/models/SleepQuality.java
@@ -6,10 +6,10 @@ import lombok.Data;
 
 @Data
 public class SleepQuality {
-    private float sleepQuality;
+    private double sleepQuality;
     private Date date;
 
-    public SleepQuality(Date date, float sleepQuality) {
+    public SleepQuality(Date date, double sleepQuality) {
         this.date = date;
         this.sleepQuality = sleepQuality;
     }
